docs: Document integrating w/ other event handlers (#249)
This PR adds a section in the Cloudflare Workers documentation on how to integrate a Hono app with other event handlers in a Module Worker.

This is based on @yusukebe's comment here: https://github.com/honojs/hono/issues/1720

I was looking to integrate a `scheduled` fn with my Hono app, and couldn't find anything in the docs for it. Hope this is useful!