update landing page copy
this pull request updates the landing page copy of hono.dev.

a lot of people were asking in Hacker News "what does this do?" / "I can't tell what this is by the website".

the updates copy here re-focus hono in the lead title/subtitle as "Hono / Web application framework", with the perks being listed underneath in the paragraph.

the "View Docs" button text was updated to "Get Started". trying to guide people more to getting started as the immediate CTA button on the homepage.