Update pylon.md: Add bun requirement disclaimer
Hey @yusukebe.

I want to add a disclaimer to figure out if the bun requirement of pylon is a dealbreaker for the community.

I am already working on a solution to support the same runtimes as Hono.

I just want to validate this feature. Could you share your personal opinion on this matter?